requirements.txt
colorama
numpy
pyalsaaudio
pyfiglet
reedsolo
termcolor
